move auth_backend and middleware to client, test views,urls,templates to admin
