+++ houdini/houdini/urls.py
@@ -3,6 +3,7 @@ from django.contrib import admin
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
+    url(r'^', include('houdini_admin.urls')),
     url(r'^', include('houdini_server.urls')),
     url(r'^', include('houdini_client.urls')),
 ]
