+++ houdini/houdini_client/views.py
@@ -15,7 +15,6 @@ import urllib.parse
 
 from houdini_server.endpoints import Endpoint
 from houdini_server.models import User
-from .decorators import login_required, role_required, permission_required
 from .forms import LoginForm, RegisterForm
 
 def login(request):
@@ -172,18 +171,3 @@ def logout(request):
         messages.error(request, r.text)
 
         return redirect('index')
-
-@login_required
-def login_test(request):
-    return render(request, "houdini_client/login_test.html")
-
-@role_required('new role')
-def role_test(request):
-    return render(request, "houdini_client/role_test.html")
-
-@permission_required('new permission')
-def permission_test(request):
-    return render(request, "houdini_client/permission_test.html")
-
-def unauthorized_401(request):
-    return render(request, "houdini_client/401.html")
